move tests to a common file
